简化pipeline Fixes #126
