+++ include/cocpp/interface/co_pipeline.h
@@ -1,5 +1,7 @@
+_Pragma("once");
 #include "cocpp/comm/co_chan.h"
 #include "cocpp/interface/co.h"
+#include "cocpp/interface/pipeline/co_pipeline_op.h"
 #include "cocpp/utils/co_noncopyable.h"
 #include <algorithm>
 #include <cstddef>
@@ -10,88 +12,6 @@
 #include <vector>
 
 CO_NAMESPACE_BEGIN
-
-namespace pipeline
-{
-
-template <typename CollectionType>
-concept Iterable = requires(CollectionType c)
-{
-    c.begin();
-    c.end();
-};
-
-template <typename FuncType, typename ItemType>
-concept PipelineInitFunc = requires(FuncType f, ItemType i)
-{
-    requires std::invocable<FuncType>;
-    requires std::same_as<std::invoke_result_t<FuncType>, std::optional<ItemType>>;
-};
-
-template <typename ItemType, typename CollectionType>
-concept To = std::is_same_v<std::decay_t<decltype(*std::declval<CollectionType>().begin())>, ItemType>;
-
-template <typename CollectionType>
-concept Collection = requires(CollectionType c)
-{
-    c.begin();
-    c.end();
-    c.insert(c.begin(), *c.begin());
-};
-
-template <typename FilterType, typename ItemType>
-concept FilterFunc = requires(FilterType f, ItemType i)
-{
-    requires std::invocable<FilterType, ItemType>;
-    requires std::same_as < std::invoke_result_t<FilterType, ItemType>,
-    bool > ;
-};
-
-template <typename ReduceType, typename InitType, typename ItemType>
-concept ReduceFunc = requires(ReduceType r, InitType i, ItemType it)
-{
-    requires std::invocable<ReduceType, InitType, ItemType>;
-    requires std::same_as<std::invoke_result_t<ReduceType, InitType, ItemType>, InitType>;
-};
-
-template <typename FuncType, typename ItemType>
-concept ReturnIsNotVoid = requires(FuncType f, ItemType i)
-{
-    requires std::invocable<FuncType, ItemType>;
-    requires !std::same_as<std::invoke_result_t<FuncType, ItemType>, void>;
-};
-
-template <typename IterType>
-concept Iterator = requires(IterType i)
-{
-    i.operator*();
-    i.operator++();
-    i.operator!=(i);
-};
-
-template <typename ArrayType>
-concept Array = requires(ArrayType a)
-{
-    {
-        std::begin(a)
-        } -> Iterator;
-    {
-        std::end(a)
-        } -> Iterator;
-};
-
-template <typename ItemType, typename IncrementType>
-concept Incrementable = requires(const ItemType& i, const IncrementType& inc)
-{
-    {
-        i + inc
-        } -> std::same_as<ItemType>;
-    {
-        i == i
-        } -> std::same_as<bool>;
-};
-
-}
 
 // 前置声明
 namespace pipeline
@@ -108,83 +28,6 @@ struct take;
 struct skip;
 template <typename FuncType>
 struct fork;
-}
-
-///////////////////////////////////////////////////////////////////////
-namespace pipeline
-{
-struct chan
-{
-};
-
-template <Collection Type>
-struct to
-{
-};
-
-template <typename FilterType>
-struct filter
-{
-    const FilterType filter__;
-    filter(const FilterType& filter)
-        : filter__(filter)
-    {
-    }
-};
-
-template <typename ReduceType, typename InitType>
-struct reduce
-{
-    const ReduceType reducer__;
-    const InitType   init__;
-    reduce(const ReduceType& reducer, const InitType& init)
-        : reducer__(reducer)
-        , init__(init)
-    {
-    }
-};
-
-struct stream
-{
-    const int size__;
-    stream(int size = -1)
-        : size__(size)
-    {
-    }
-};
-
-struct take
-{
-    const size_t take__;
-    take(const size_t& tk)
-        : take__(tk)
-    {
-    }
-};
-
-struct skip
-{
-    size_t skip__;
-    skip(const size_t& sp)
-        : skip__(sp)
-    {
-    }
-};
-
-template <typename FuncType>
-struct fork
-{
-    size_t   fork_count__;
-    FuncType func__;
-    fork(const size_t& fc, const FuncType& func)
-        : fork_count__(fc)
-        , func__(func)
-    {
-    }
-};
-
-///////////////////////////////////////////////////////////////////////
-
 }
 
 ////////////////////////////////////////////////////////////////////////////
